requirements.txt
atomicwrites==1.4.0
bidict==0.22.0
black==19.10b0
blinker==1.5
Brotli==1.0.9
brotlipy==0.7.0
cachelib==0.9.0
chart-studio==1.1.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.1
Flask-Assets==2.0
Flask-Compress==1.12
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SeaSurf==1.1.1
Flask-Session==0.4.0
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
inflection==0.5.1
itsdangerous==2.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.23.1
pandas==1.4.3
plotly==5.9.0
ply==3.11
pyls-spyder==0.4.0
PyQt5==5.15.7
python-dotenv==0.20.0
python-engineio==4.3.4
python-lsp-jsonrpc==1.0.0
python-socketio==5.7.2
PyYAML==6.0
SQLAlchemy==1.4.41
tenacity==8.0.1
ua-parser==0.16.1
webassets==2.0
webencodings==0.5.1
Werkzeug==2.2.1
wincertstore==0.2
WTForms==3.0.1